# backend/core/docs/swagger.yaml
basePath: /api/v1
definitions:
  articles.ArticleInput:
    properties:
      authorname:
        type: string
      content:
        type: string
      draft:
        type: boolean
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
    required:
    - authorname
    - content
    - subtitle
    - title
    type: object
...
  articles.EditArticleInput:
    properties:
      content:
        type: string
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
    required:
    - content
    - subtitle
    - title
    type: object
  articles.GetArticlesTopicRespond:
    properties:
      "true":
        example: TestTopic
        type: string
    type: object
...
  articles.IsArticleDraftResponse:
    properties:
      "true":
        example: Article is draft
        type: string
    type: object
  articles.LikesResponse:
    properties:
      accounts:
        items:
          type: integer
        type: array
      amount:
        type: integer
    type: object
...
  articles.MultipleArticlesIdsInput:
    type: object
...
  http.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  main_router_articles.Article:
    properties:
      authorName:
        type: string
      content:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      likes:
        items:
          type: integer
        type: array
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
      userId:
        type: integer
    type: object
  main_router_articles.DeletedResponse:
    properties:
      delete:
        example: all articles have been successfully deleted
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /articles:
    delete:
      consumes:
      - application/json
      description: Delete all the connected user articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.DeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Retrieve all articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Add an article
      parameters:
      - description: Params to create an article
        in: body
        name: ArticleInput
        required: true
        schema:
          $ref: '#/definitions/articles.ArticleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Edit an article
      parameters:
      - description: Params to edit an article
        in: body
        name: EditArticleInput
        required: true
        schema:
          $ref: '#/definitions/articles.EditArticleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/:id:
    delete:
      consumes:
      - application/json
      description: Delete an article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.DeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Retrieve an article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/:id/draft:
    get:
      consumes:
      - application/json
      description: Get all topics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.IsArticleDraftResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Change the state of a draft
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/:id/likes:
    delete:
      consumes:
      - application/json
      description: Remove a like from an article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Retrieve likes information related to an article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.LikesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Add a like to an article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/:id/topic:
    get:
      consumes:
      - application/json
      description: Get the topic of a Article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.GetArticlesTopicRespond'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/latest/created:
    get:
      consumes:
      - application/json
      description: Retrived articles created in the last two hours
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/latest/modified:
    get:
      consumes:
      - application/json
      description: Retrived articles modified in the last two hours
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/liked:
    get:
      consumes:
      - application/json
      description: Retrieve liked articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/me:
    get:
      consumes:
      - application/json
      description: Retrieve connected user articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/multiples:
    post:
      consumes:
      - application/json
      description: Get Articles by there Ids
      parameters:
      - description: "\tIds of the Articles research"
        in: body
        name: MultipleArticlesIdsInput
        required: true
        schema:
          $ref: '#/definitions/articles.MultipleArticlesIdsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/search/:keyword:
    post:
      consumes:
      - application/json
      description: Find Articles by a keyword and search if it correspond
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/topic/:topic:
    get:
      consumes:
      - application/json
      description: Get all articles by topic
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
  /articles/topics:
    get:
      consumes:
      - application/json
      description: Get all articles by topic
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"